+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "body": "^5.1.0",
     "body-parser": "^1.20.2",
+    "dotenv": "^16.3.1",
     "ejs": "^3.1.9",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "mysql2": "^3.5.2"
   }
 }
