+++ utils/messages.js
@@ -1,0 +1,1 @@
+/* 상수 메시지 파일 미리 생성함 */
